requirements.txt
jsonfield

Django==3.1.2

Pillow==7.2.0

pytesseract==0.3.7


